requirements.txt
amqp==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
cachetools==5.3.2
celery==5.4.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cron-descriptor==1.4.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-braces==1.15.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-import-export==3.3.6
django-js-asset==2.0.0
django-markdown==0.8.4
django-misaka==0.2.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Faker==17.3.0
fonttools==4.51.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
houdini.py==0.1.0
idna==3.6
ipython==8.21.0
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.7
Markdown==3.5.2
MarkupPy==1.14
matplotlib==3.8.4
matplotlib-inline==0.1.6
misaka==2.1.1
module-name==0.6.0
mysql==0.0.3
mysqlclient==2.2.0
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.4
parso==0.8.3
Pillow==9.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
pyotp==2.9.0
pyparsing==3.1.2
python-crontab==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
soupsieve==2.4.1
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
stack-data==0.6.3
tablib==3.5.0
threadpoolctl==3.2.0
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
